get functionality working, add Day of Week, add singleExperienceList
